feat(user_profile): add golden tests and refactor widgets for testability
- Extract `ProfileHeader`, `ProfileSection`, and `ProfileInfoTile` from `UserProfilePage` to standalone widgets.
- Add golden tests for `ProfileHeader` and `ProfileSection`.
- Update `UserProfilePage` golden tests to cover Loaded, Loading, and Error states.
- Generate reference images for all new and updated golden tests.
- Remove redundant golden test file.
